check existing tract maps

+++ maps.py
@@ -4,8 +4,8 @@ import json
 import polyline
 import urllib.request
 
-conn = psycopg2.connect(user='geogeist', password='password',
-                        host='localhost', port='5432')
+conn = psycopg2.connect(user='geogeist', password=os.environ['DBPASS'],
+						host='localhost', port='5432')
 
 MAP_KEY = os.environ['MAP_KEY']
 URL_MAX = 8192
@@ -21,20 +21,24 @@ def tracts():
 		state = row[3]
 		data_json = row[4]
 		gid = row[5]
-		# TODO: handle multipolygons
-		coords = geog["coordinates"][0][0]
-		for item in coords:
-			item.reverse()
-		poly = polyline.encode(coords)
-
-		url = "https://maps.googleapis.com/maps/api/staticmap?size=400x400&center=%s,%s&zoom=14&path=%senc:%s&key=%s" % (centroid[1],centroid[0],"fillcolor:0xAA000033%7Ccolor:0xFFFFFF00%7C",poly,MAP_KEY)
 		filename = "static/tract_map_%s-%s.png" % (state, name)
-		urllib.request.urlretrieve(url, filename)
+		if os.path.exists(filename):
+			print('map already exists: ' + filename)
+		else:
+			# TODO: handle multipolygons
+			coords = geog["coordinates"][0][0]
+			for item in coords:
+				item.reverse()
+			poly = polyline.encode(coords)
 
-		data_json['map'] = filename
+			url = "https://maps.googleapis.com/maps/api/staticmap?size=400x400&center=%s,%s&zoom=14&path=%senc:%s&key=%s" % (centroid[1],centroid[0],"fillcolor:0xAA000033%7Ccolor:0xFFFFFF00%7C",poly,MAP_KEY)
 
-		print(state + "-" + name)
-		cur.execute("UPDATE tracts SET data = %s WHERE gid = %s", (json.dumps(data_json), gid))
+			urllib.request.urlretrieve(url, filename)
+
+			data_json['map'] = filename
+
+			print("rendered: " + filename)
+			cur.execute("UPDATE tracts SET data = %s WHERE gid = %s", (json.dumps(data_json), gid))
 	conn.commit()
 	cur.close()
 
@@ -119,5 +123,5 @@ def counties():
 	cur.close()
 
 
-counties()
+tracts()
 conn.close()
